[feh-background] Fix loosing focus
On wallpaper change there was a nasty bug, when window focus moved to
the window under cursor. Now there's a workaround with `xdotool` (which
is a new dependency btw), and a shortcut, if no compton is running.

P.S. Just now wallpaper changed and I didn't had to fix focus while I
write this commit message. Luv it so much)