[i3] Fix default directory
The daemonization was done as described usually, with chdir to /. But in
fact, it is *suggested* to go to root. The only requirements are for the
workdir to exist as long, as daemon lasts. And if my homedir goes
somewhere, I'm not shure I will care about stashed terminal emulator)