[feh-background] Fix error with no active window
It's been been a while since I ran this script. When I did - there was a
weird case when i3wm leaves screen w/o active window, when stashed
terminal should be active and IS in focus... Weird.

Nontheless, this change accounts this case and preserves weird errors in
logs.