[qutebrowser] Add global theme
It was annonced previously. Now it works (yay) and I add it to the repo.
Due to lack of CSS-conditions support, this has to work via UserScripts.
The css2js.sh script converts user.css to user.js, and adds nessesary
checks and features.

This theme works *somewhat* good on most sites, but still far from
stable.